refactor: move remote_config code to the root of dd-trace (#4475)
