+++ packages/dd-trace/src/appsec/rule_manager.js
@@ -2,7 +2,7 @@
 
 const fs = require('fs')
 const waf = require('./waf')
-const { ACKNOWLEDGED, ERROR } = require('./remote_config/apply_states')
+const { ACKNOWLEDGED, ERROR } = require('../remote_config/apply_states')
 
 const blocking = require('./blocking')
