+++ packages/dd-trace/src/remote_config/manager.js
@@ -3,13 +3,13 @@
 const { URL, format } = require('url')
 const uuid = require('crypto-randomuuid')
 const { EventEmitter } = require('events')
-const tracerVersion = require('../../../../../package.json').version
-const request = require('../../exporters/common/request')
-const log = require('../../log')
-const { getExtraServices } = require('../../service-naming/extra-services')
+const tracerVersion = require('../../../../package.json').version
+const request = require('../exporters/common/request')
+const log = require('../log')
+const { getExtraServices } = require('../service-naming/extra-services')
 const { UNACKNOWLEDGED, ACKNOWLEDGED, ERROR } = require('./apply_states')
 const Scheduler = require('./scheduler')
-const { GIT_REPOSITORY_URL, GIT_COMMIT_SHA } = require('../../plugins/util/tags')
+const { GIT_REPOSITORY_URL, GIT_COMMIT_SHA } = require('../plugins/util/tags')
 
 const clientId = uuid()
